agrego lista de endpoints

+++ src/main/java/com/gestion/SNYA/repositorio/InstitucionRepositorio.java
@@ -1,13 +1,15 @@
 package com.gestion.SNYA.repositorio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
 
 import com.gestion.SNYA.modelo.Institucion;
 
 @Repository
 
-public interface InstitucionRepositorio extends JpaRepository<Institucion, Long> {
+public interface InstitucionRepositorio extends JpaRepository<Institucion, Integer> {
+	
 	
 
 }
